Pulled out hybrid motion model tests
